Adding common props into CodeContainer
