+++ src/components/Tools/EncoderDecoder.jsx/EncoderCode.jsx
@@ -1,16 +1,5 @@
 import React from "react";
-import styled from "styled-components";
-import { toast } from "react-toastify";
 import { CodeContainer } from "./EncoderCodeElement";
-
-const CopyButton = styled.button`
-    padding: 5px 10px;
-    background-color: #ff6b08;
-    color: white;
-    border: none;
-    border-radius: 3px;
-    cursor: pointer;
-`;
 
 function EncoderCode(props) {
     function Base64() {
@@ -60,43 +49,13 @@ function EncoderCode(props) {
         });
     }
 
-    function copyToClipboard(text) {
-        navigator.clipboard.writeText(text).then(() => {
-            toast.success("Copied to clipboard!");
-        });
-    }
-
     return (
         <div>
-            <h1>Base64 Encode</h1>
-            <CodeContainer>
-                <p>{Base64()}</p>
-                <CopyButton onClick={() => copyToClipboard(Base64())}>Copy</CopyButton>
-            </CodeContainer>
-
-            <h1>Url Encode</h1>
-            <CodeContainer>
-                <p>{Url()}</p>
-                <CopyButton onClick={() => copyToClipboard(Url())}>Copy</CopyButton>
-            </CodeContainer>
-
-            <h1>Full Url Encode</h1>
-            <CodeContainer>
-                <p>{fullURLEncode(props.Input)}</p>
-                <CopyButton onClick={() => copyToClipboard(fullURLEncode(props.Input))}>Copy</CopyButton>
-            </CodeContainer>
-
-            <h1>ASCII HEX Encode</h1>
-            <CodeContainer>
-                <p>{asciiHexEncode(props.Input)}</p>
-                <CopyButton onClick={() => copyToClipboard(asciiHexEncode(props.Input))}>Copy</CopyButton>
-            </CodeContainer>
-
-            <h1>HTML Encode</h1>
-            <CodeContainer>
-                <p>{htmlEncode(props.Input)}</p>
-                <CopyButton onClick={() => copyToClipboard(htmlEncode(props.Input))}>Copy</CopyButton>
-            </CodeContainer>
+            <CodeContainer text={Base64()} title={"Base64 Encode"} />
+            <CodeContainer text={Url()} title={"Url Encode"} />
+            <CodeContainer text={fullURLEncode(props.Input)} title={"Full Url Encode"} />
+            <CodeContainer text={asciiHexEncode(props.Input)} title={"ASCII HEX Encode"} />
+            <CodeContainer text={htmlEncode(props.Input)} title={"HTML Encode"} />
         </div>
     );
 }
