+++ src/components/Tools/EncoderDecoder.jsx/EncoderCodeElement.jsx
@@ -1,6 +1,8 @@
 import styled from "styled-components";
+import React from "react";
+import { toast } from "react-toastify";
 
-export const CodeContainer = styled.div`
+export const Container = styled.div`
     height: 50px;
     background: #131313;
     margin-bottom: 10px;
@@ -13,3 +15,30 @@ export const CodeContainer = styled.div`
     overflow: clip;
     color: white;
 `;
+
+const CopyButton = styled.button`
+    padding: 5px 10px;
+    background-color: #ff6b08;
+    color: white;
+    border: none;
+    border-radius: 3px;
+    cursor: pointer;
+`;
+
+export const CodeContainer = ({ text, title }) => {
+    function copyToClipboard(text) {
+        navigator.clipboard.writeText(text).then(() => {
+            toast.success("Copied to clipboard!");
+        });
+    }
+
+    return (
+        <>
+            <h1>{title}</h1>
+            <Container>
+                <p>{text}</p>
+                <CopyButton onClick={() => copyToClipboard(text)}>Copy</CopyButton>
+            </Container>
+        </>
+    );
+};
